Move ProxyConsts.cs to Utils directory
