+++ RyuSocks/Utils/ProxyConsts.cs
@@ -14,7 +14,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-namespace RyuSocks
+namespace RyuSocks.Utils
 {
     public static class ProxyConsts
     {
